Some Updates
